UISE-97: Move data to models in Search component and Filters. Mode Search component to separate folder. Delete not used data from filters config
