+++ src/Filters/index.js
@@ -1,0 +1,2 @@
+export { default as Filters } from './Filters';
+export * from './model';
