+++ src/Search/model/availableIndexes.js
@@ -1,0 +1,12 @@
+export default [
+  { label: 'Search all fields', value: '', localOnly: true },
+  { label: 'FOLIO ID', value: 'id' },
+  { label: 'Title', value: 'title' },
+  { label: 'Identifier', value: 'identifier' },
+  { label: 'ISBN', value: 'identifier/type=isbn' },
+  { label: 'ISSN', value: 'identifier/type=issn' },
+  { label: 'Contributor', value: 'contributor', localOnly: true },
+  { label: 'Subject', value: 'subject' },
+  { label: 'Classification', value: 'classification', localOnly: true },
+  { label: 'Publisher', value: 'publisher' },
+];
