+++ src/Search/model/index.js
@@ -1,0 +1,4 @@
+export { default as availableIndexes } from './availableIndexes';
+export { default as filterConfig } from './filterConfig';
+export { default as parseFiltersString } from './parseFiltersString';
+export { default as redirectParams } from './redirectParams';
